# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shop API",
    "description": "Невеличка тестова API магазину товарів (Products, Users, Orders)",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "paths": {
    "/products": {
      "get": {
        "summary": "Отримати список товарів",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "Список товарів",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Створити товар",
        "tags": ["Products"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductInput" } } }
        },
        "responses": {
          "201": {
            "description": "Товар створено",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/products/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/IdParam" }],
      "get": {
        "summary": "Отримати товар за id",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "Товар",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "summary": "Оновити товар",
        "tags": ["Products"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductInput" } } }
        },
        "responses": {
          "200": {
            "description": "Товар оновлено",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "summary": "Видалити товар",
        "tags": ["Products"],
        "responses": {
          "204": { "description": "Товар видалено" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Отримати список користувачів",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "Список користувачів",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Створити користувача",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserInput" } } }
        },
        "responses": {
          "201": {
            "description": "Користувача створено",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/users/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/IdParam" }],
      "get": {
        "summary": "Отримати користувача за id",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "Користувач",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "summary": "Оновити користувача",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserInput" } } }
        },
        "responses": {
          "200": {
            "description": "Користувача оновлено",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "summary": "Видалити користувача",
        "tags": ["Users"],
        "responses": {
          "204": { "description": "Користувача видалено" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Отримати список замовлень",
        "tags": ["Orders"],
        "responses": {
          "200": {
            "description": "Список замовлень",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Order" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Створити замовлення",
        "tags": ["Orders"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderInput" } } }
        },
        "responses": {
          "201": {
            "description": "Замовлення створено",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/orders/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/IdParam" }],
      "get": {
        "summary": "Отримати замовлення за id",
        "tags": ["Orders"],
        "responses": {
          "200": {
            "description": "Замовлення",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "summary": "Оновити замовлення",
        "tags": ["Orders"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Замовлення оновлено",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "summary": "Видалити замовлення",
        "tags": ["Orders"],
        "responses": {
          "204": { "description": "Замовлення видалено" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "IdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Не знайдено",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "BadRequest": {
        "description": "Невалідні дані запиту",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "price": { "type": "number" },
          "stock": { "type": "integer" },
          "category": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true }
        }
      },
      "ProductInput": {
        "type": "object",
        "required": ["name", "price"],
        "properties": {
          "name": { "type": "string" },
          "price": { "type": "number" },
          "stock": { "type": "integer" },
          "category": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" }
        }
      },
      "UserInput": {
        "type": "object",
        "required": ["name", "email"],
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "userId": { "type": "integer" },
          "productId": { "type": "integer" },
          "quantity": { "type": "integer" },
          "status": { "type": "string", "enum": ["pending", "completed", "cancelled"] }
        }
      },
      "OrderInput": {
        "type": "object",
        "required": ["userId", "productId", "quantity"],
        "properties": {
          "userId": { "type": "integer" },
          "productId": { "type": "integer" },
          "quantity": { "type": "integer" }
        }
      },
      "OrderUpdate": {
        "type": "object",
        "properties": {
          "quantity": { "type": "integer" },
          "status": { "type": "string", "enum": ["pending", "completed", "cancelled"] }
        }
      }
    }
  }
}